no longer calls delay functions.
each call to TM1637Display::update() will perform the next step
of the data transmission to the display (68 total for full write)
if there is one to be performed. each call to the same will return
a bool which is false if further calls to update() are needed. if
update() returns true all transmission is completed and another may
be initiated.